+++ src/views/admin/formula/components/CreateFormula.tsx
@@ -1,7 +1,6 @@
-import { useEffect, useState } from 'react'
-import { showModal } from '../../../../stores/formula/formulaSlice'
+import { showModal } from 'stores/formula/formulaSlice'
 import { useAppDispatch, useAppSelector } from 'stores/hooks'
-import { RootState } from '../../../../stores/'
+import { RootState } from 'stores/'
 
 // custom components
 import Card from 'components/card/Card'
@@ -18,61 +17,9 @@ import {
 
 
 const CreateFormula = () => {
-  const calculation = useAppSelector((state:RootState) => state.formula.calculation)
+  const result = useAppSelector((state:RootState) => state.formula.result)
   const dispatch = useAppDispatch()
   const textColor = useColorModeValue("navy.700", "white");
-
-  const [result, setResult] = useState<number>(0)
-
-  const doCalculation = (calculation: string) => {
-      let arr = calculation.split('')
-      let num = ''
-  
-      for(let i = 0; i < calculation.length; i++) {
-        let char = calculation.charAt(i)
-        switch (char) {
-          case '+':
-          case 'x':
-          case '-':
-          case '/':
-            arr.push(num)
-            arr.push(char)
-            num = ''
-            break
-          default:
-            num += char
-            break
-        }
-      }
-      arr.push(num)
-  
-      let hasil = parseInt(arr[0])
-      for(let i = 1; i < arr.length; i += 2) {
-        switch (arr[i]) {
-          case '+':
-            hasil += parseInt(arr[i+1])
-            break
-          case 'x':
-            hasil *= parseInt(arr[i+1])
-            break
-          case '-':
-            hasil -= parseInt(arr[i+1])
-            break
-          case '/':
-            hasil /= parseInt(arr[i+1])
-            break
-        }
-      }
-      if(!hasil){
-        return 0
-      }
-      return hasil
-    }
-
-  useEffect(() => {
-    setResult(doCalculation(calculation))
-  }, [calculation])
-  
 
   return (
     <Card p='20px'>
